Set CARGO_HOME
This PR ensures that CARGO_HOME is set for subsequenty buildpacks at build time. CARGO_HOME is set to a unique layer, which is empty at the end of this buildpack execution.

This enables tools to locate CARGO_HOME automatically, since CARGO_HOME is a well-known env variables used by Rust tools.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>