Mark BOM entry as Launch
While Rust is not a launch dependency, it's version is important to know because it statically compiles binaries. Thus we are going to include its dependency into the launch BOM. This also works around an issue with buildpacks where the build BOM is not really visible. It's not included in `pack inspect` and the BOM label on the image.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>